# docs/api.yaml
openapi: "3.0.3"

info:
  version: 1.0.0
  title: Xclone-API
  description: 練習用Xclone-API

servers:
  - url: http://localhost:3000
    description: "開発者用サーバー"

paths:
  /tweets:
    get:
      summary: 全ツイート取得
      tags:
        - Tweet
      responses:
        200:
          description: success operation
          content:
            application/json:
              schema:
                type: "array"
                example:
                  - id: "ef304778ada8a6"
                    author: "ネコニキ"
                    title: "ReactNative@v0.63で追加されたPressableが地味にすごい"
                    body: "ReactNative@v0.63で追加されたPressableが地味にすごい"
                    link: "https://zenn.dev/nekoniki/articles/ef304778ada8a6"
        404:
          description: "NotFound"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: "string"
                    example: "NOT_FOUND"
                  message:
                    type: "string"
                    example: "NOT_FOUND"
    post:
      summary: ツイート作成
      tags:
        - Tweet
      requestBody:
        description: 新規ツイートを送信する。
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - author
                - title
                - body
                - link
              properties:
                id:
                  type: string
                  example: "ef304778ada8a6"
                author:
                  type: string
                  example: "ネコニキ"
                title:
                  type: string
                  example: "ReactNative@v0.63で追加されたPressableが地味にすごい"
                body:
                  type: string
                  example: "ReactNative@v0.63で追加されたPressableが地味にすごい"
                link:
                  type: string
                  example: "https://zenn.dev/nekoniki/articles/ef304778ada8a6"
      responses:
        201:
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  success:
                    type: "string"
                    example: "REGISTERED"
        400:
          description: "BadRequest"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: string
                    example: BAD_REQUEST
                  message:
                    type: string
                    example: "BAD_REQUEST"

  /tweets/{id}:
    get:
      summary: 指定したtweetの詳細を取得API
      tags:
        - Tweet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: success operation
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  id:
                    type: string
                    example: "ef304778ada8a6"
                  author:
                    type: string
                    example: "ネコニキ"
                  title:
                    type: string
                    example: "ReactNative@v0.63で追加されたPressableが地味にすごい"
                  body:
                    type: string
                    example: "ReactNative@v0.63で追加されたPressableが地味にすごい"
                  link:
                    type: string
                    example: "https://zenn.dev/nekoniki/articles/ef304778ada8a6"
        404:
          description: "NotFound"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: "string"
                    example: "NOT_FOUND"
                  message:
                    type: "string"
                    example: "NOT_FOUND"
    delete:
      summary: 指定したtweetの削除
      tags:
        - Tweet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "Success"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  success:
                    type: "string"
                    example: "DELETED"
        400:
          description: "BadRequest"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: string
                    example: BAD_REQUEST
                  message:
                    type: string
                    example: "BAD_REQUEST"
        403:
          description: "Forbidden"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: string
                    example: "FORBIDDEN"
                  message:
                    type: string
                    example: "FORBIDDEN"
        404:
          description: "NotFound"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: "string"
                    example: "NOT_FOUND"
                  message:
                    type: "string"
                    example: "NOT_FOUND"